requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
jinja2
pdfkit
python-multipart
email-validator
